Move Table to legacy
