+++ frictionless/helpers.py
@@ -18,6 +18,9 @@ from _thread import RLock  # type: ignore
 from . import config
 
 
+# TODO: remove unused
+
+
 # General
