+++ tests/legacy/test_table.py
